as/azf/requirements.txt
azure-functions
azure-functions-durable
azure-identity
requests
pydantic
pytest
requests-mock
